objects | choosing object in statement

+++ bots/applicant/statement.py
@@ -1,5 +1,11 @@
 from app.models import Statement
-from app.services import statementservice, notificationservice, supplierservice, apiservice
+from app.services import (
+    statementservice, 
+    notificationservice, 
+    supplierservice, 
+    apiservice, 
+    objectservice
+    )
 from . import *
 from bots import *
 from bots.applicant.directive import to_the_typing_product_name
@@ -17,7 +23,8 @@ def _to_the_typing_product_amount(update, order):
 
 def _to_the_typing_product_comment(update, order):
     text = get_word('type product comment', update) + ' ({})'.format(order.product)
-    update_message_reply_text(update, text)
+    markup = reply_keyboard_markup([[get_word('back', update)]])
+    update_message_reply_text(update, text, markup)
     return GET_PRODUCT_COMMENT
 
 def _to_the_asking_action(update, obj):
@@ -31,16 +38,49 @@ def _to_the_asking_action(update, obj):
         text_order += '\n➖➖➖➖➖➖➖\n'
         text += text_order
     
-    text += '\n' + get_word('end statement or add more', update)
+    text += '\n' + get_word('continue or add more', update)
     reply_markup = reply_keyboard_markup(
         [
             [get_word('add product', update)], 
-            [get_word('end statement', update)]
+            [get_word('continue', update)],
+            [get_word('back', update)],
+
         ]
         )
 
     update_message_reply_text(update, text, reply_markup)
     return GET_ACTION
+
+def _to_the_typing_object(update):
+    text = get_word('type object', update)
+    reply_markup = get_objects_reply_markup()
+    reply_markup.keyboard.append([KeyboardButton(text=get_word('back', update))])
+    update_message_reply_text(update, text, reply_markup)
+    return GET_OBJECT
+
+def _to_the_finish_statement(update, obj):
+    text = get_word('your order', update) + ':\n\n'
+    orders = statementservice.filter_orders_of_object(obj)
+    for order in orders:
+        text_order = get_word('order details', update)
+        text_order = text_order.format(
+            title=order.product, amount=order.amount, product_comment=order.comment
+        )
+        text_order += '\n➖➖➖➖➖➖➖\n'
+        text += text_order
+    
+    text += '\n{}: <b>{}</b>'.format(get_word('object', update), obj.object.title)
+    reply_markup = reply_keyboard_markup(
+        [
+            [get_word('end statement', update)], 
+            [get_word('back', update)],
+
+        ]
+        )
+
+    update_message_reply_text(update, text, reply_markup)
+    return FINISH_STATEMENT
+
 
 def _end_statement(update, context, obj):
     
@@ -66,15 +106,18 @@ def get_product_name(update, context):
     if msg == get_word('back', update):
         # check statement has products
         if statementservice.filter_unfinished_objects_by_update(update):
-            # delete last order
-            statementservice.remove_order_from_obj(obj)
-            return _to_the_asking_action(update, obj)
+            # check, there are 2 or more orders in statement, otherwise go to main menu
+            if len(statementservice.filter_orders_of_object(obj)) > 1:
+            
+                # delete last order
+                statementservice.remove_order_from_obj(obj)
+                return _to_the_asking_action(update, obj)
 
         bot_delete_message(
             update, context, 
             update.message.message_id-1, 
         )
-
+        delete_unfinished_statements(update)
         main_menu(update, context)
         return ConversationHandler.END
     
@@ -126,10 +169,41 @@ def get_action(update, context):
     msg = update.message.text
     obj = statementservice.get_current_object_by_update(update)
     order = statementservice.get_last_order_of_object(obj)
-    if msg == get_word('add product', update):
+    if msg == get_word('back', update):
+        return _to_the_typing_product_comment(update, order)
+    
+    elif msg == get_word('add product', update):
         
         # add new order to statement
         statementservice.add_order_to_the_obj(obj)
         return to_the_typing_product_name(update, context)
+    elif msg == get_word('continue', update):
+        return _to_the_typing_object(update)
+
+@is_start
+def get_object(update, context):
+    msg = update.message.text
+    obj = statementservice.get_current_object_by_update(update)
+    order = statementservice.get_last_order_of_object(obj)
+    if msg == get_word('back', update):
+        return _to_the_asking_action(update, obj)
+
+    if objectservice.is_msg_object_title(msg):
+        object = objectservice.get_object_by_msg(msg)
+        obj.object = object
+        obj.save()
+        return _to_the_finish_statement(update, obj)  
+    else:
+        update_message_reply_text(update, get_word('choose specified objects', update))
+        return
+
+@is_start
+def finish_statement(update, context):
+    msg = update.message.text
+    obj = statementservice.get_current_object_by_update(update)
+    order = statementservice.get_last_order_of_object(obj)
+    if msg == get_word('back', update):
+        return _to_the_typing_object(update)
+
     elif msg == get_word('end statement', update):
         return _end_statement(update, context, obj)
